feat: famille metier de la bcn et fix insersup opendata (#383)
* feat: famille de métier de la bcn

* docs: mise à jour des documentations sur les familles de métiers

* fix: continuum diplome inexistant

* fix: stats regionales

* feat: nouveau millesime

* feat: url temporaire insersup

* fix: insersup but inexistant

* fix: insersup opendata

* docs: maj doc api bcn

* docs: maj schema dat

* fix: ajout des nature_uai manquante

* feat: remove new millesime

+++ server/src/jobs/stats/importCertificationsSupStats.js
@@ -58,6 +58,11 @@ export async function importCertificationsSupStats(options = {}) {
               const stats = omitNil(pick(formationStats, ["millesime", ...INSERSUP_STATS_NAMES]));
               const customStats = computeCustomStats(stats, "insersup", INSERSUP_CUSTOM_STATS_NAMES);
 
+              if (!certification) {
+                handleError(new Error(`Diplome ${formationStats.code_certification} inconnu`));
+                return null;
+              }
+
               // Delete data compute with continuum job (= when type is not self)
               await certificationsStats().deleteOne({
                 ...query,
